chore(ci): dedupe report resolution and delta formatting
Two cleanups from a quality pass over the ci-report stack:
- Extract the PR-vs-base report resolution (sha-suffixed report, plain
  file as baseline only when its sha differs) into a shared
  ci-report/report-files.mjs, replacing the ~40-line verbatim copy in
  both the bundle-size and eager-graph section writers.
- Drop the eager-graph writer's local formatDelta clone; add a noBaseline
  option to the shared formatDelta so the one function covers the
  "no base measurement" wording it needed.

No behavior change — the consolidated comment renders identically.

Generated-By: PostHog Code
Task-Id: 47cda15c-e2d1-43f3-acca-31b57bbb90be
